Enabled specification of multi-part test cases in gui with single test case number; Fixed hybridization, file download, and file criteria test cases.

+++ qa/software/API_Test_Suite/src/caarray/client/test/suite/SearchByCriteriaTestSuite.java
@@ -58,9 +58,9 @@ public abstract class SearchByCriteriaTestSuite extends ConfigurableTestSuite
                 {
                     search.setTestCase(Float.parseFloat(input[headerIndexMap.get(TEST_CASE)]
                                                               .trim()));
-                    if (!excludeTests.isEmpty() && (excludeTests.contains(search.getTestCase()) || excludeTests.contains(Math.floor(search.getTestCase()))))
+                    if (!excludeTests.isEmpty() && (excludeTests.contains(search.getTestCase()) || excludeTests.contains((float)Math.floor(search.getTestCase()))))
                         skip = true;
-                    if (!includeTests.isEmpty() && (!includeTests.contains(search.getTestCase())&& !includeTests.contains(Math.floor(search.getTestCase()))))
+                    if (!includeTests.isEmpty() && (!includeTests.contains(search.getTestCase())&& !includeTests.contains((float)Math.floor(search.getTestCase()))))
                         skip = true;
                 }
                 if (!skip)
@@ -83,9 +83,9 @@ public abstract class SearchByCriteriaTestSuite extends ConfigurableTestSuite
                 if (search == null)
                     throw new TestConfigurationException(
                             "No test case indicated for row: " + index);
-                if (!excludeTests.isEmpty() && (excludeTests.contains(search.getTestCase()) || excludeTests.contains(Math.floor(search.getTestCase()))))
+                if (!excludeTests.isEmpty() && (excludeTests.contains(search.getTestCase()) || excludeTests.contains((float)Math.floor(search.getTestCase()))))
                     skip = true;
-                if (!includeTests.isEmpty() && (!includeTests.contains(search.getTestCase())&& !includeTests.contains(Math.floor(search.getTestCase()))))
+                if (!includeTests.isEmpty() && (!includeTests.contains(search.getTestCase())&& !includeTests.contains((float)Math.floor(search.getTestCase()))))
                     skip = true;
                 if (!skip)
                 {try
@@ -191,7 +191,7 @@ public abstract class SearchByCriteriaTestSuite extends ConfigurableTestSuite
             List<CriteriaSearch> filteredSearches = new ArrayList<CriteriaSearch>();
             for (CriteriaSearch search : configuredSearches)
             {
-                if (includeTests.contains(search.getTestCase()) || includeTests.contains(Math.floor(search.getTestCase())))
+                if (includeTests.contains(search.getTestCase()) || includeTests.contains((float)Math.floor(search.getTestCase())))
                     filteredSearches.add(search);
             }
             configuredSearches = filteredSearches;
